# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: Wigope Pay API
  version: 1.0.0
  description: Production API contract for Wigope Pay consumer app, provider integrations, admin operations, and webhooks.
servers:
  - url: https://api.wigope.com/api/v1
    description: Production
  - url: http://127.0.0.1:4000/api/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /recharge/detect-operator:
    get:
      summary: Detect mobile operator
      parameters:
        - in: query
          name: number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operator detection result
  /recharge/mobile:
    post:
      summary: Initiate prepaid recharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number, amount]
              properties:
                number:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Recharge status
...
  /transactions:
    get:
      summary: All transaction history
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
      responses:
        '200':
          description: Transactions
  /transactions/{id}:
    get:
      summary: Transaction detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction detail
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    hubbleSecret:
      type: apiKey
      in: header
      name: X-Hubble-Secret
    kwikapiSignature:
      type: apiKey
      in: header
      name: x-kwikapi-signature
...